Move value::types to crate root
