+++ artichoke-backend/src/convert/array.rs
@@ -2,7 +2,7 @@ use std::convert::TryFrom;
 
 use crate::convert::{Convert, Error, Int, TryConvert};
 use crate::sys;
-use crate::value::types::{Ruby, Rust};
+use crate::types::{Ruby, Rust};
 use crate::value::Value;
 use crate::Artichoke;
