+++ artichoke-backend/src/convert/nilable/array.rs
@@ -1,7 +1,7 @@
 use crate::convert::fixnum::Int;
 use crate::convert::float::Float;
 use crate::convert::{Convert, Error, TryConvert};
-use crate::value::types::{Ruby, Rust};
+use crate::types::{Ruby, Rust};
 use crate::value::Value;
 use crate::Artichoke;
 
@@ -186,7 +186,7 @@ mod tests {
     use crate::convert::{Convert, TryConvert};
     use crate::eval::Eval;
     use crate::sys;
-    use crate::value::types::Ruby;
+    use crate::types::Ruby;
     use crate::value::Value;
 
     #[test]
