+++ artichoke-backend/src/convert/nilable/fixnum.rs
@@ -1,6 +1,6 @@
 use crate::convert::fixnum::Int;
 use crate::convert::{Convert, Error, TryConvert};
-use crate::value::types::{Ruby, Rust};
+use crate::types::{Ruby, Rust};
 use crate::value::Value;
 use crate::Artichoke;
 
@@ -44,7 +44,7 @@ mod tests {
     use crate::convert::{Convert, TryConvert};
     use crate::eval::Eval;
     use crate::sys;
-    use crate::value::types::Ruby;
+    use crate::types::Ruby;
     use crate::value::Value;
 
     #[test]
