+++ artichoke-backend/src/convert/string.rs
@@ -1,5 +1,5 @@
 use crate::convert::{Convert, Error, TryConvert};
-use crate::value::types::{Ruby, Rust};
+use crate::types::{Ruby, Rust};
 use crate::value::Value;
 use crate::Artichoke;
 
@@ -60,7 +60,7 @@ mod tests {
     use crate::convert::{Convert, Error, TryConvert};
     use crate::eval::Eval;
     use crate::sys;
-    use crate::value::types::{Ruby, Rust};
+    use crate::types::{Ruby, Rust};
     use crate::value::Value;
 
     #[test]
