+++ artichoke-backend/src/lib.rs
@@ -229,6 +229,7 @@ pub mod method;
 pub mod module;
 pub mod state;
 pub mod top_self;
+pub mod types;
 pub mod value;
 pub mod warn;
 
@@ -261,9 +262,9 @@ pub enum ArtichokeError {
     /// Failed to create an [argspec](sys::args) `CString`.
     ArgSpec,
     /// Failed to convert from a Rust type to a [`sys::mrb_value`].
-    ConvertToRuby(convert::Error<value::types::Rust, value::types::Ruby>),
+    ConvertToRuby(convert::Error<types::Rust, types::Ruby>),
     /// Failed to convert from a [`sys::mrb_value`] to a Rust type.
-    ConvertToRust(convert::Error<value::types::Ruby, value::types::Rust>),
+    ConvertToRust(convert::Error<types::Ruby, types::Rust>),
     /// Exception raised during eval.
     ///
     /// See [`Eval`](eval::Eval).
